check authorize when deleting

+++ src/main/java/com/example/prj1be/controller/BoardController.java
@@ -41,7 +41,11 @@ public class BoardController {
   }
 
   @DeleteMapping("remove/{id}")
-  public ResponseEntity remove(@PathVariable Integer id) {
+  public ResponseEntity remove(@PathVariable Integer id, @SessionAttribute(value = "login", required = false) Member login) {
+    if (login == null) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
+    if (!service.hasAccess(id, login)) return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
     if (service.remove(id)) return ResponseEntity.ok().build();
     else return ResponseEntity.badRequest().build();
   }
